Moved command line parsing code to new file
